front-endSPA/authentication/auth/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg==3.2.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
sqlparse==0.5.1
typing_extensions==4.12.2
Pillow==9.5.0
